Split scripts, improve confirmation window, implement messaging between processes
