+++ package.json
@@ -15,10 +15,11 @@
         "prod": {}
     },
     "devDependencies": {
-        "@parcel/config-webextension": "^2.0.0-beta.3.1",
-        "@parcel/runtime-browser-hmr": "^2.0.0-beta.3.1",
-        "@parcel/transformer-image": "^2.0.0-beta.3.1",
-        "@parcel/transformer-sass": "^2.0.0-beta.3.1",
+        "@parcel/config-webextension": "^2.4.1",
+        "@parcel/runtime-browser-hmr": "^2.4.1",
+        "@parcel/transformer-image": "^2.4.1",
+        "@parcel/transformer-raw": "^2.4.1",
+        "@parcel/transformer-sass": "^2.4.1",
         "@types/chrome": "^0.0.124",
         "@types/firefox": "^0.0.30",
         "@types/firefox-webext-browser": "^82.0.0",
@@ -27,6 +28,7 @@
         "@types/react": "^17.0.2",
         "@types/react-dom": "^17.0.2",
         "@types/react-test-renderer": "^17.0.1",
+        "@types/webextension-polyfill": "^0.8.3",
         "@typescript-eslint/eslint-plugin": "^4.4.1",
         "@typescript-eslint/parser": "^4.4.1",
         "eslint": "^7.11.0",
@@ -37,7 +39,7 @@
         "jest": "^27.0.5",
         "jest-css-modules": "^2.1.0",
         "jest-webextension-mock": "^3.7.14",
-        "parcel": "^2.0.0-beta.3.1",
+        "parcel": "^2.4.1",
         "prettier": "^2.1.2",
         "react-test-renderer": "^17.0.2",
         "ts-jest": "^27.0.3",
@@ -57,7 +59,8 @@
         "build": "npm run cleanup && parcel build src/manifest.json --target prod",
         "cleanup": "rm -rf dist .cache .parcel-cache",
         "test": "jest --config=jest.config.js",
-        "lint": "eslint --fix -c ./.eslintrc.js \"src/**/*.ts*\"",
+        "lint": "eslint -c ./.eslintrc.js \"src/**/*.ts*\"",
+        "fix": "eslint --fix -c ./.eslintrc.js \"src/**/*.ts*\"",
         "prettify": "prettier --write \"src/**/*.ts*\""
     }
 }
