+++ src/confirmation-window/styles.css
@@ -1,11 +1,8 @@
 .confirmation-container {
-    width: 400px;
-    height: 560px;
     overflow-y: auto;
     background-color: white;
     font-family: Arial, sans-serif;
     padding: 20px;
-    border: 2px solid black;
 }
 
 .confirmation-title {
@@ -40,6 +37,7 @@
     padding: 10px 20px;
     font-size: 16px;
     border: none;
+    cursor: pointer;
 }
 
 .button.button-primary {
