--- src/context/reducers/index.ts
@@ -1,4 +1,0 @@
-import { ManageViewReducer } from "pointsdk/context/reducers/manageViewReducer";
-import { WalletReducer } from "pointsdk/context/reducers/walletReducer";
-
-export { ManageViewReducer, WalletReducer };
