+++ src/popup/component.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { browser } from "webextension-polyfill-ts";
-import { Toolbar } from "pointsdk/components/toolbar";
-import { Navigation } from "pointsdk/components/navigation";
-import { Apps } from "pointsdk/components/apps";
-import { AppProvider } from "pointsdk/context";
+import { Toolbar } from "pointsdk/popup/components/toolbar";
+import { Navigation } from "pointsdk/popup/components/navigation";
+import { Apps } from "pointsdk/popup/components/apps";
+import { AppProvider } from "pointsdk/popup/context";
 import "./style.scss";
 
 export const Popup: React.FC = () => {
