+++ src/popup/components/apps/__tests__/apps.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { Apps } from "../component";
-import { Story } from "pointsdk/components/dev";
+import { Story } from "pointsdk/popup/components/dev";
 
 storiesOf("Apps", module).add("renders", () => {
     return (
