+++ src/popup/components/apps/style.scss
