+++ src/popup/components/navigation/__tests__/navigation.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { Navigation } from "../component";
-import { Story } from "pointsdk/components/dev";
+import { Story } from "pointsdk/popup/components/dev";
 
 storiesOf("Navigation", module).add("renders", () => {
     return (
