+++ src/popup/components/navigation/style.scss
