+++ src/popup/components/toolbar/__tests__/toolbar.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { Toolbar } from "../component";
-import { Story } from "pointsdk/components/dev";
+import { Story } from "pointsdk/popup/components/dev";
 
 storiesOf("Toolbar", module).add("renders", () => {
     return (
