+++ src/popup/components/toolbar/style.scss
