+++ src/popup/components/wallet/__tests__/wallet.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { Wallet } from "../component";
-import { Story } from "pointsdk/components/dev";
+import { Story } from "pointsdk/popup/components/dev";
 
 storiesOf("Wallet", module).add("renders", () => {
     return (
