+++ src/popup/components/wallet/component.tsx
