+++ src/popup/components/wallet/index.tsx
