+++ src/popup/components/wallet/style.scss
