+++ src/popup/context/reducers/index.ts
@@ -1,0 +1,4 @@
+import { ManageViewReducer } from "pointsdk/popup/context/reducers/manageViewReducer";
+import { WalletReducer } from "pointsdk/popup/context/reducers/walletReducer";
+
+export { ManageViewReducer, WalletReducer };
