+++ src/popup/context/reducers/manageViewReducer.ts
