+++ src/popup/index.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="utf-8"/>
         <title>Point Network Extension</title>
-        <script src="popup/index.tsx"></script>
+        <script src="index.tsx"></script>
     </head>
     <body><div id="popup"></div></body>
 </html>
